Move new react components to components/options/, following latest changes
