+++ src/components/options/UnsubmittedVideosComponent.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import Config from "../config";
+import Config from "../../config";
 import UnsubmittedVideoListComponent from "./UnsubmittedVideoListComponent";
 
 export interface UnsubmittedVideosProps {
